Renamed some Buffer methods to better reflect the SSH datatype

+++ src/test/java/net/schmizz/sshj/util/BufferTest.java
@@ -59,7 +59,7 @@ public class BufferTest {
         assertEquals(handyBuf.putString("some string").readString(), "some string");
 
         // uint32
-        assertEquals(handyBuf.putInt(0xffffffffL).readLong(), 0xffffffffL);
+        assertEquals(handyBuf.putUInt32(0xffffffffL).readUInt32(), 0xffffffffL);
     }
 
     @Test
